refactor: split DocsTab and ws-handler hotspots by domain
Extract collection WS handlers into a dedicated module and pull DocsTab
query/menu/import-export types into components/docs for clearer ownership
without behavior changes.
