+++ packages/client/src/components/docs/docsQuery.test.ts
@@ -1,0 +1,63 @@
+import { describe, expect, it } from "vitest";
+import type { DocSummary } from "../../store/types";
+import {
+	buildQueryChips,
+	matchesQuery,
+	parseQuery,
+	relativeTime,
+	stripToken,
+} from "./docsQuery";
+
+function doc(partial: Partial<DocSummary> & { name: string }): DocSummary {
+	return {
+		id: partial.name,
+		category: "general",
+		format: "A4",
+		pageCount: 1,
+		elementCount: 0,
+		...partial,
+	};
+}
+
+describe("docsQuery", () => {
+	it("parses multi-criteria tokens", () => {
+		expect(parseQuery("@flyer #locked :4 summer")).toEqual({
+			category: "flyer",
+			locked: true,
+			minRating: 4,
+			text: "summer",
+		});
+	});
+
+	it("matches docs against query filters", () => {
+		const flyer = doc({
+			name: "Summer promo",
+			category: "flyer",
+			locked: true,
+			rating: 5,
+		});
+		const q = parseQuery("@flyer #locked :4 summer");
+		expect(matchesQuery(flyer, q)).toBe(true);
+		expect(
+			matchesQuery(doc({ name: "Summer promo", category: "poster" }), q),
+		).toBe(false);
+	});
+
+	it("strips tokens and builds removable chips", () => {
+		const search = "@flyer #locked :3";
+		const query = parseQuery(search);
+		const chips = buildQueryChips(query, search, () => {});
+		expect(chips.map((c) => c.key)).toEqual(["cat", "lock", "rating"]);
+		expect(stripToken(search, (t) => t === "#locked")).toBe("@flyer :3");
+	});
+
+	it("formats relative time in fr and en", () => {
+		const ts = new Date(Date.now() - thrMinutes(5)).toISOString();
+		expect(relativeTime(ts, "en")).toMatch(/m ago|just now/);
+		expect(relativeTime(ts, "fr")).toMatch(/min|instant/);
+	});
+});
+
+function thrMinutes(n: number): number {
+	return n * 60_000;
+}
